Split off benchmark/__init__.py

Remove duplicated code in __init__.py
